修复。

+++ src/com/pengrad/telegrambot/model/User.java
@@ -9,14 +9,14 @@ import java.io.Serializable;
 public class User implements Serializable {
     private final static long serialVersionUID = 0L;
 
-    private Integer id;
+    private Long id;
     private Boolean is_bot;
     private String first_name;
     private String last_name;
     private String username;
     private String language_code;
 
-    public Integer id() {
+    public Long id() {
         return id;
     }
